update: sticker 관련  서비스 로직 분리
