+++ functions/controller/userContoller.js
@@ -42,8 +42,6 @@ module.exports = {
       slackAPI.sendMessageToSlack(slackMessage, slackAPI.DEV_WEB_HOOK_ERROR_MONITORING);
 
       res.status(statusCode.INTERNAL_SERVER_ERROR).send(util.fail(statusCode.INTERNAL_SERVER_ERROR, responseMessage.INTERNAL_SERVER_ERROR));
-    } finally {
-      client.release();
     }
   },
 };
